pyproject.toml
[tool.poetry]
name = "new-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "new_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django-rest-framework = "^0.1.0"
pillow = "^9.4.0"
django = "^4.1.7"
pyjwt = "^2.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pyright = "^1.1.298"
isort = "^5.12.0"
black = "^23.1.0"
pytest-django = "^4.5.2"
taskipy = "^1.10.3"
django-stubs = "^1.15.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.2"
django-mkdocs = "^0.1.2"
mkdocstrings = "^0.20.0"
mkdocstrings-python = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"
DJANGO_SETTINGS_MODULE = "core.settings"


[tool.taskipy.tasks]
lint = "black --diff . && isort --diff ."
coverage = "coverage html"
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=core --cov=menu --cov=user -vv"
post_test = "task coverage"
runserver = "python manage.py makemigrations && python manage.py migrate && python manage.py runserver"
